connect with redux
